add Database.Validate() and reuse for seed validation
Test coverage is provided by the existing seed validation tests.
This introduces yet more code in the short term, but the error reporting
structure is designed such that most or all of the UI validation code
can be replaced with it later. This makes all current and future
validation code accessible to unit tests without having to involve the
UI in it.