switch from sha256_crypt to libxcrypt
This adds support for newer and stronger password hashes like bcrypt and
yescrypt, depending on the libxcrypt version.

For added cryptoagility, we will automatically rehash old passwords
using stronger methods when the user logs into the UI. This is the only
opportunity where the password is presented to Portunus itself (and not
just the LDAP), so the user has to use specifically the Portunus UI.

For seeded users, rehashing is done whenever the seed is reapplied, i.e.
in practice, immediately upon starting up and loading the DB from disk.

Closes #2.