remove all frontend validation where redundant with the core validation
This requires an extensive rework of all form interactions. We now
submit all CRUD forms to Nexus.Update() in order to collect the errors
from the core validation. Some custom validation rules remain in the
frontend that cannot be represented in the core (errors arising from
type conversion and password input handling). If a custom validation
rule has already failed, the Nexus.Update() is executed as a dry run
solely to collect additional errors for other fields.