fix a deadlock in the LDAP adapter initialization
For large databases, the initial immediate call of the AddListener
callback would exceed the buffer size of `operationsChan`. Since this
channel only starts being drained after AddListener has returned, this
resulted in a deadlock.

I'm moving this to the same design as for the store adapter for
consistency, and since a buffer size of 1 is easier to understand than
the previous arbitrarily selected bound.