fix missing r.ParseForm() call to populate r.PostForm
From the "how did this ever work" department.

I could not login on my development setup running at
<http://localhost:8080/>, and this was reproducible on v2.1.4, even
though login definitely works on that version running in production.

My best guess is that some middleware that I removed in the previous
commits was calling r.ParseForm() as a side-effect, in a way that did
not happen when not behind TLS or something.